Moved a few things
