+++ src/ProudNet/ProudNet.csproj
@@ -86,7 +86,7 @@
     <Compile Include="Exceptions.cs" />
     <Compile Include="Extensions.cs" />
     <Compile Include="ISessionFactory.cs" />
-    <Compile Include="P2PGroup.cs" />
+    <Compile Include="P2PConnectionState.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Configuration.cs" />
     <Compile Include="ProudSession.cs" />
@@ -113,14 +113,13 @@
     <Compile Include="Serialization\Serializers\StreamSerializer.cs" />
     <Compile Include="Serialization\Serializers\StreamWithScalarSerializer.cs" />
     <Compile Include="Serialization\Serializers\StringSerializer.cs" />
-    <Compile Include="Server\P2PGroup.cs" />
-    <Compile Include="Server\P2PGroupManager.cs" />
-    <Compile Include="Server\ProudServer.cs" />
-    <Compile Include="Server\RemotePeer.cs" />
-    <Compile Include="Server\Handlers\SessionHandler.cs" />
-    <Compile Include="Server\Handlers\CoreServerHandler.cs" />
+    <Compile Include="P2PGroup.cs" />
+    <Compile Include="P2PGroupManager.cs" />
+    <Compile Include="ProudServer.cs" />
+    <Compile Include="RemotePeer.cs" />
+    <Compile Include="Handlers\SessionHandler.cs" />
     <Compile Include="Handlers\ProudMessageHandler.cs" />
-    <Compile Include="Server\Handlers\ServerHandler.cs" />
+    <Compile Include="Handlers\ServerHandler.cs" />
     <Compile Include="Handlers\CoreHandler.cs" />
   </ItemGroup>
   <ItemGroup>
